devtestlabs: consolidating to a single Virtual Machine ID parser
